+++ scripts/hologram_modem.py
@@ -38,13 +38,31 @@ help_signal = '''This subcommand prints the RSSI signal strength values.\n
 help_location = '''This subcommand prints the encoded location of the modem.\n
 '''
 
+help_reset = '''This subcommand causes the modem to restart itself.\n
+'''
+
+help_radio_off = '''This subcommand causes the modem to turn off the cellular
+radio\n
+'''
+
+help_radio_on = '''This subcommand causes the modem to turn on the cellular
+radio\n
+'''
+
 def run_modem_connect(args):
+    print 'Note: "hologram modem connect" is deprecated '\
+            'in favor of "hologram network connect"'
     cloud = CustomCloud(None, network='cellular')
     cloud.network.disable_at_sockets_mode()
-    cloud.network.connect()
-    print 'PPP session started'
+    res = cloud.network.connect()
+    if res:
+        print 'PPP session started'
+    else:
+        print 'Failed to start PPP'
 
 def run_modem_disconnect(args):
+    print 'Note: "hologram modem disconnect" is deprecated '\
+            'in favor of "hologram network disconnect"'
     print 'Checking for existing PPP sessions'
     for proc in psutil.process_iter():
 
@@ -67,6 +85,28 @@ def run_modem_signal(args):
             handle_timeout(args['repeat'])
     else:
         print 'Signal strength: ' + str(cloud.network.signal_strength)
+
+def run_modem_reset(args):
+    cloud = CustomCloud(None, network='cellular')
+    cloud.network.modem.reset()
+    print 'Restarted modem'
+
+def run_modem_radio_off(args):
+    cloud = CustomCloud(None, network='cellular')
+    res = cloud.network.modem.radio_power(False)
+    if res:
+        print 'Modem radio disabled'
+    else:
+        print 'Failure to disable radio'
+
+def run_modem_radio_on(args):
+    cloud = CustomCloud(None, network='cellular')
+    res = cloud.network.modem.radio_power(True)
+    if res:
+        print 'Modem radio enabled'
+    else:
+        print 'Failure to enable radio'
+
 
 def run_modem_sim(args):
     cloud = CustomCloud(None, network='cellular')
@@ -95,7 +135,10 @@ _run_handlers = {
     'modem_operator': run_modem_operator,
     'modem_signal': run_modem_signal,
     'modem_type': run_modem_type,
-    'modem_location': run_modem_location
+    'modem_location': run_modem_location,
+    'modem_reset': run_modem_reset,
+    'modem_radio_on': run_modem_radio_on,
+    'modem_radio_off': run_modem_radio_off
 }
 
 # EFFECTS: Parses the CLI arguments as options to the hologram modem subcommand.
@@ -139,6 +182,22 @@ def parse_hologram_modem_args(parser):
     parser_location = subparsers.add_parser('location', help=help_location)
     parser_location.set_defaults(command_selected='modem_location')
     parser_location.add_argument('-v', nargs='?', action=VAction, dest='verbose', required=False)
+
+    # Reset
+    parser_reset = subparsers.add_parser('reset', help=help_reset)
+    parser_reset.set_defaults(command_selected='modem_reset')
+    parser_reset.add_argument('-v', nargs='?', action=VAction, dest='verbose', required=False)
+
+    # radio-on
+    parser_radio_on = subparsers.add_parser('radio-on', help=help_reset)
+    parser_radio_on.set_defaults(command_selected='modem_radio_on')
+    parser_radio_on.add_argument('-v', nargs='?', action=VAction, dest='verbose', required=False)
+
+    # radio-off
+    parser_radio_off = subparsers.add_parser('radio-off', help=help_reset)
+    parser_radio_off.set_defaults(command_selected='modem_radio_off')
+    parser_radio_off.add_argument('-v', nargs='?', action=VAction, dest='verbose', required=False)
+
 
 # EFFECTS: Runs the hologram modem interfaces.
 def run_hologram_modem(args):